// FilmLibraryTest.Web/ClientApp/src/apiClient/apiClient.ts
export interface IResultModelOfFilm {
    error?: string | undefined;
    result?: Film | undefined;
}

export interface Film {
    id: number;
    title?: string | undefined;
    description?: string | undefined;
    year: number;
    producer?: string | undefined;
    posterUrl?: string | undefined;
    user?: User | undefined;
}
// ...
export interface IdentityUser extends IdentityUserOfString {
}

export interface User extends IdentityUser {
    tokens?: Token[] | undefined;
    films?: Film[] | undefined;
}

export interface Token {
    id: number;
    value?: string | undefined;
    user?: User | undefined;
}